New deferred question behaviour where students can explain their answer
In other respects, this works like the deferred feedback behaviour. The
explantion is not graded automatically, not is it mandatory to give one.
However, the teacher may choose to manually adjust the marks based on the
reasons (e.g. to give 'method marks' for wrong answers.) Alternatively,
this coudl be a way for students to be able to reflect on what they were
thinking when the attempted the question, when they grades are released.